pyproject.toml
[project]
name = "hookgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=4.3.1",
    "ffmpeg-python>=0.2.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-google-genai>=2.0.10",
    "numpy>=2.1.3",
    "opencv-python-headless>=4.11.0.86",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "retina-face>=0.0.17",
    "streamlit>=1.46.1",
    "tf-keras>=2.19.0",
    "torch>=2.7.1",
    "webrtcvad-wheels>=2.0.14",
    "yt-dlp>=2025.6.9",
]
